increased test coverage of DataFrame / DataMatrix, miscellaneous fixes
git-svn-id: http://pandas.googlecode.com/svn/trunk@100 d5231056-7de3-11de-ac95-d976489f1ece

+++ pandas/core/panel.py
@@ -1044,7 +1044,7 @@ class LongPanel(Panel):
 
         I, N, K = self.dims
 
-        values = np.empty((I, N, K), dtype=float)
+        values = np.empty((I, N, K), dtype=self.values.dtype)
 
         mask = self.index.mask
         notmask = -mask
